+++ src/Pgvector.EntityFrameworkCore/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.2.1 (2024-06-25)
+
+- Added support for `halfvec` and `sparsevec` types
+- Added support for compiled models
+- Added `L1Distance`, `HammingDistance`, and `JaccardDistance` functions
+
+## 0.2.0 (2023-11-24)
+
+- Added support for Npgsql.EntityFrameworkCore.PostgreSQL 8
+- Dropped support for Npgsql.EntityFrameworkCore.PostgreSQL < 8
+
+## 0.1.2 (2023-09-25)
+
+- Added distance functions
+
 ## 0.1.1 (2023-04-25)
 
 - Added support for .NET 6